refactor(local-mode): move local mode config to `./local-mode.ts`
To keep the same file structure as in in `./dev-mode.ts`.
