+++ core/src/plugins/kubernetes/config.ts
@@ -23,15 +23,13 @@ import {
   containerArtifactSchema,
   ContainerEnvVars,
   containerEnvVarsSchema,
-  containerLocalModeSchema,
-  ContainerLocalModeSpec,
   ContainerRegistryConfig,
   containerRegistryConfigSchema,
 } from "../container/moduleConfig"
 import { PluginContext } from "../../plugin-context"
 import { dedent, deline } from "../../util/string"
 import { defaultSystemNamespace } from "./system"
-import { syncableKinds, SyncableKind } from "./types"
+import { SyncableKind, syncableKinds } from "./types"
 import { BaseTaskSpec, baseTaskSpecSchema, cacheResultSchema } from "../../config/task"
 import { BaseTestSpec, baseTestSpecSchema } from "../../config/test"
 import { ArtifactSpec } from "../../config/validation"
@@ -39,39 +37,8 @@ import { V1Toleration } from "@kubernetes/client-node"
 import { runPodSpecIncludeFields } from "./run"
 import { DevModeDefaults, devModeDefaultsSchema } from "./dev-mode"
 import { KUBECTL_DEFAULT_TIMEOUT } from "./kubectl"
-import { localModeGuideLink } from "./local-mode"
 
 export const DEFAULT_KANIKO_IMAGE = "gcr.io/kaniko-project/executor:v1.8.1-debug"
-
-export interface KubernetesLocalModeSpec extends ContainerLocalModeSpec {
-  containerName?: string
-  target?: KubernetesTargetResourceSpec
-}
-
-export const kubernetesLocalModeSchema = () =>
-  containerLocalModeSchema().keys({
-    containerName: joi
-      .string()
-      .optional()
-      .description(
-        "When using the `defaultTarget` and not specifying `localMode.target`, this field can be used to override the default container name to proxy traffic from."
-      ),
-    target: targetResourceSpecSchema().description(
-      "The remote Kubernetes resource to proxy traffic from. If specified, this is used instead of `defaultTarget`."
-    ),
-  }).description(dedent`
-    Configures the local application which will send and receive network requests instead of the target resource specified by \`localMode.target\` or \`defaultTarget\`. One of those fields must be specified to enable local mode for the action.
-
-    The selected container of the target Kubernetes resource will be replaced by a proxy container which runs an SSH server to proxy requests.
-    Reverse port-forwarding will be automatically configured to route traffic to the locally run application and back.
-
-    Local mode is enabled by setting the \`--local\` option on the \`garden deploy\` or \`garden dev\` commands.
-    Local mode always takes the precedence over dev mode if there are any conflicting service names.
-
-    Health checks are disabled for services running in local mode.
-
-    See the [Local Mode guide](${localModeGuideLink}) for more information.
-  `)
 
 export interface ProviderSecretRef {
   name: string
